feat: add contract rent details and yield calculations for rented listings

+++ api/listings/[id].ts
@@ -74,7 +74,7 @@ export default async function handler(req: VercelRequest, res: VercelResponse) {
               l.is_rented, l.lifetime_rent, ${isFavoriteSelect}, ${isHiddenSelect}, l.active,
               l.inactive_since, l.first_seen, l.last_seen,
               re.estimated_rent, re.avg_rent_per_m2, re.sample_count, re.confidence, re.match_level, re.rental_yield,
-              rcd.current_rent AS rent_current_rent, rcd.contract_expiry_date AS rent_contract_expiry
+              rcd.current_rent::float AS rent_current_rent, rcd.contract_expiry_date AS rent_contract_expiry
        FROM listings l
        ${joinClause}
        LEFT JOIN rental_estimates re ON re.listing_id = l.id
